+++ examples/generative/corrdiff/generate.py
@@ -39,7 +39,7 @@ from modulus.utils.generative import (
     parse_int_list,
     StackedRandomGenerator,
 )
-from module import Module  # TODO import from Core once the kwargs issue is fixed
+from modulus import Module
 
 from datasets.base import DownscalingDataset
 from datasets.dataset import init_dataset_from_config
@@ -63,11 +63,12 @@ def main(cfg: DictConfig) -> None:
     class_idx = getattr(cfg, "class_idx", None)  # TODO: is this needed?
     num_steps = getattr(cfg, "num_steps", 18)
     sample_res = getattr(cfg, "sample_res", "full")
-    res_edm = getattr(cfg, "res_edm", True)
     sampling_method = getattr(cfg, "sampling_method", "stochastic")
     seed_batch_size = getattr(cfg, "seed_batch_size", 1)
     force_fp16 = getattr(cfg, "force_fp16", False)
     use_torch_compile = getattr(cfg, "use_torch_compile", True)
+    regression_only = getattr(cfg, "regression_only", False)
+    diffusion_only = getattr(cfg, "diffusion_only", False)
 
     # Parse deterministic sampler options
     sigma_min = getattr(cfg, "sigma_min", None)
@@ -138,6 +139,7 @@ def main(cfg: DictConfig) -> None:
     # Initialize distributed manager
     DistributedManager.initialize()
     dist = DistributedManager()
+    device = dist.device
 
     # Initialize logger
     logger = PythonLogger("generate")  # General python logger
@@ -159,30 +161,31 @@ def main(cfg: DictConfig) -> None:
 
     logger0.info(f"torch.__version__: {torch.__version__}")
 
-    # Load diffusion network
-    logger0.info(f'Loading residual network from "{res_ckpt_filename}"...')
-    net_res = Module.from_checkpoint(res_ckpt_filename)
-
-    # load regression network
-    if res_edm:
-        logger0.info(f'Loading network from "{reg_ckpt_filename}"...')
-        net_reg = Module.from_checkpoint(reg_ckpt_filename)
-    else:
+    # Sanity check for the type of requested inference
+    if regression_only and diffusion_only:
+        raise ValueError(
+            "Both regression_only and diffusion_only cannot be set to True."
+        )
+    if regression_only:
+        net_res = None
+    if diffusion_only:
         net_reg = None
 
-    # move to device
-    device = dist.device
-    net_res = net_res.eval().to(device).to(memory_format=torch.channels_last)
-    net_reg = (
-        net_reg.eval().to(device).to(memory_format=torch.channels_last)
-        if net_reg
-        else None
-    )
+    # Load diffusion network, move to device, change precision
+    if not regression_only:
+        logger0.info(f'Loading residual network from "{res_ckpt_filename}"...')
+        net_res = Module.from_checkpoint(res_ckpt_filename)
+        net_res = net_res.eval().to(device).to(memory_format=torch.channels_last)
+        if force_fp16:
+            net_res.use_fp16 = True
 
-    # change precision if needed
-    if force_fp16:
-        net_reg.use_fp16 = True
-        net_res.use_fp16 = True
+    # load regression network, move to device, change precision
+    if not diffusion_only:
+        logger0.info(f'Loading network from "{reg_ckpt_filename}"...')
+        net_reg = Module.from_checkpoint(reg_ckpt_filename)
+        net_reg = net_reg.eval().to(device).to(memory_format=torch.channels_last)
+        if force_fp16:
+            net_reg.use_fp16 = True
 
     # Reset since we are using a different mode.
     if use_torch_compile:
@@ -190,7 +193,8 @@ def main(cfg: DictConfig) -> None:
         compile_mode = "reduce-overhead"
         # Only compile residual network
         # Overhead of compiling regression network outweights any benefits
-        net_res = torch.compile(net_res, mode=compile_mode)
+        if net_res:
+            net_res = torch.compile(net_res, mode=compile_mode)
 
     def generate_fn(image_lr):
         """Function to generate an image
@@ -220,7 +224,7 @@ def main(cfg: DictConfig) -> None:
             logger0.info(f"seeds: {sample_seeds}")
             if net_reg:
                 with nvtx.annotate("regression_model", color="yellow"):
-                    image_mean = generate(
+                    image_reg = generate(
                         net=net_reg,
                         img_lr=image_lr_patch,
                         seed_batch_size=1,
@@ -230,9 +234,9 @@ def main(cfg: DictConfig) -> None:
                         pretext="reg",
                         class_idx=class_idx,
                     )
-
+            if net_res:
                 with nvtx.annotate("diffusion model", color="purple"):
-                    image_out = image_mean + generate(
+                    image_res = generate(
                         net=net_res,
                         img_lr=image_lr_patch.expand(seed_batch_size, -1, -1, -1).to(
                             memory_format=torch.channels_last
@@ -245,20 +249,12 @@ def main(cfg: DictConfig) -> None:
                         num_steps=num_steps,
                         **sampler_kwargs,
                     )
-
+            if regression_only:
+                image_out = image_reg
+            elif diffusion_only:
+                image_out = image_res
             else:
-                with nvtx.annotate("diffusion model", color="purple"):
-                    image_out = generate(
-                        net=net_res,
-                        img_lr=image_lr_patch.expand(seed_batch_size, -1, -1, -1),
-                        seed_batch_size=seed_batch_size,
-                        sampling_method=sampling_method,
-                        seeds=sample_seeds,
-                        pretext="gen",
-                        class_idx=class_idx,
-                        num_steps=num_steps,
-                        **sampler_kwargs,
-                    )
+                image_out = image_reg + image_res
 
             # reshape: (1*9*9)x3x50x50  --> 1x3x450x450
             if sample_res != "full":